Increase character prefab movement speed
The prototype character prefab's movement was a bit slow. Since we
are going to be using the ninja sprite, I feel like movement should
be more fast to fit that theme. The minimum speed was sped up from 1f
to 5f so that it would feel fast but still relatively easy to control
when wanting to maneuver more precise jumps. The maximum speed was
increased so that it would feel more fast paced but not fast enough
so that you run into walls a lot. Acceleration was reduced to give
a bit more feeling of momentum but it's not that noticeable.