Update starter prefab that spawns first in level
The starting prefab chunk is now Chunkdesign1-1 since it was the easiest
to work with. It needed boxCollider2D for each ground block and crate so
I added those to each in ONLY that prefab. They all are used by
composite collider.

This was done so that I can see if future chunks are spawning in
correctly.