Fix implementation for increasing speed over time
Fix the way that speed is increased. The previous implementation
was incorrect in the assumptions made as to what acceleration was
refering to. I made the mistake thinking that acceleration refered
to a speed multiplier when it was in fact about the force being
applied to make the character acheive min or max speed.

The implementation in this commit is still not as good as it could be.
I think that in the future we should opt for a linear interpolation
that takes abs(x) as the input so that its a continuous and smooth
increase rather than a discrete and hard increment. But this can be
changed in a later sprint if we so desire it.