Apply boxColliders to chunk prefab components
Apply boxCollider2D to all children of each new chunk prefab so that
the character wont fall through it. Used by the composite collider of
that chunk prefab.

This was a quick fix and should likely be better handled by making a
prefab for each component like the crate and long-tile pieces.