Spawn difficult chunks after a certain interval
Hard chunks will spawn after a certain interval determined by
hardChunkFreq. There is a separate array for these hard chunks than
the regular level chunks. The frequency of spawning increases over
time so thats what increases difficulty over time.

The speed of the character prefab was also increased just to make
the game a bit easier to test out.

I should note that the game play is a bit slow and more testing will
need to be done to see what values should be tweaked in the prefab.
But that is a simple variable change in the Unity UI.