Adjust character movement and chunk spawning
Adjust character movement, hard chunk spawning frequency, and speedup
around the "Chunkdesign3-*" chunks. Character speed and acceleration
was adjusted to make those chunks easier to clear. Speedup and hard
chunk spawning frequency was increased so that full speedup is acheived
quicker and hard chunks spawn more frequently over time.

Chunkdesign3-* chunks are a bit difficult to maneuver and have to be
very precise. If the game feels slow and we need to make the game feel
faster then something has to be done about this family of hard chunks.
Keep in mind for later.